Add new OID alias type regrole
The new type has the scope of whole the database cluster so it doesn't
behave the same as the existing OID alias types which have database
scope,
concerning object dependency. To avoid confusion constants of the new
type are prohibited from appearing where dependencies are made involving
it.

Also, add a note to the docs about possible MVCC violation and
optimization issues, which are general over the all reg* types.

Kyotaro Horiguchi

-- OPENGPDB additions.

I have omitted catalog changes provides within this commit for ABI compat sake.
This is not disallows use of them. One, if need, could use them with help of
extension.

Kirill Reshke