Fix wrong results in TVF scans in plans produced by ORCA after merge
The 9.4 merge introduced a new way to represent TVF RTEs in GPDB - using
RangeTblFunctions (instead of FuncExpr used earlier). This commit
correctly populates RangeTblFunctions::funcparams when translation an
ORCA generated plan.
This param list is needed when TVF contains an outer reference in its
arguments. If it is left empty, ExecReScanFunctionScan() ignores all
chgParams, and rewinds instead of resetting its tuplestores. This will
produce wrong results because the FunctionScan will continue to return
the same result (as of its first execution) for every Rescan, even when
there are outer param changes.

See attached test for an example.