Fix algebrization of subqueries in queries with complex GROUP BYs
ORCA's algebrizer must first normalize GROUP BYs in a query to a form
usable in ORCA. It must flatten expressions in the project list to
contain only aggregates and grouping columns
For example:

ORGINAL QUERY:
  SELECT * from r where r.a > (SELECT max(c) + min(d)
                               FROM t where r.b = t.e)
NEW QUERY:
  SELECT * from r where r.a > (SELECT x1+x2 as x3
  FROM (SELECT max(c) as x2, min(d) as x2
        FROM t where r.b = t.e) t2)

However this process did not support subqueries in the target list that
may contain outer references, sometimes in other (nested) subqueries. It
also did not support CTEs. All these would produce a normalization error
and fall back.

This commit fixes that by supporting subqueries & CTEs.

It also includes some refactors in related areas:

- Rename IncrLevelsUpInVar to IncrLevelsUpIfOuterRef, to capture it's
  implementation.
- Remove SContextHavingQualMutator, after realizing that it has almost
  the same members as SContextGrpbyPlMutator.
- Use MakeVarInDerivedTable & RunGroupByProjListMutator in
  RunGroupByProjListMutator to reduce code drift.
- Merge RunGroupByProjListMutator & RunHavingQualMutator (see below)

RunGroupByProjListMutator() was implemented at a later date than the
RunHavingQualMutator, and did not handle subqueries and ctes correctly.
After understanding its purpose, I think the functionality of both the
above methods should be exactly the same, since they're trying to
achieve the same goal.

So, this commit just merges the two functions together into a new
function - RunExtractAggregatesMutator. In this process, I also
discovered a bug in the old RunHavingQualMutator, that is now fixed:

  create table fooh1 (a int, b int, c int);
  insert into fooh1 select i%4, i%3, i from generate_series(1,20) i;
  select f1.a + 1 from fooh1 f1 group by f1.a+1 having sum(f1.a+1) + 1 > 20;

Finally, the earlier code deduplicated AGGREFs when possible for HAVING
clauses, but not for GROUP BY target lists when moving them into the
derived query. Not only is that inconsistent, but may also give
incorrect results in case of volatile functions. The executor already
handles de-duplicating AGGREFs in right circumstances, so doing this in
the algebrizer doesn't provide much of a benefit.