Don't copy kfw DLLs on Windows in the copylibs make target
Related PR discussion:

kmacoskey:

Another piece of this can be removed with the same intention as the
other removal.

This line in the Makefile is related to a kfw snowflake dependency that
was being vendored into the Clients/Loaders package as brought in by
Ivy. Note the path, ext. It was dropped from Ivy by releng awhile ago,
2c6f4c9#diff-271f398f0c07147024e7980737cddb7d . The cp here probably
should have been removed along with the change to Ivy, but it was
missed. It can be removed - but lets make sure the Clients team is
aware.

pf-qiu:

I think it's safe to do this now.

(cherry-picked from 22646c108fee8f6abe35e4d1fa75b17bfd1a99d8)

Co-authored-by: Bradford D. Boyle <bboyle@pivotal.io>
Co-authored-by: Ben Christel <bchristel@pivotal.io>