Replace List with RBTree for preassigned Oids
pg_upgrade can potentially preassign quite a large number of Oids
when upgrading, so the List would be a bottleneck during lookups
of Oids allocated to non-preassigned objects. This replaces the
list with a Red-Black tree to achieve O(log n) lookups at the cost
of slower inserts.

Backported from GPDB main (7X):
https://github.com/greenplum-db/gpdb/commit/62e87534feafd5e6711e4bb5de6cb1d97c10f4df