CI: fix Concourse template for gpexpand
gpexpand requires centos6, ICW, and CLI jobs to be enabled, but it was
being included in the groups when only a subset of those requirements
was met, so fly refused the pipeline. With this change, centos6/CLI
pipelines should work again.

(cherry picked from commit 52a2176783c405dea1afcd5496df1448c06b4432)