Fix gpfdist pipe failure with latest windows build
Behavior of stat() is not stable across C Runtime versions.
Implementation of msvcrt.dll calls FindFirstFile(), while in normal
CRT that omes with Visual Studio or Redistribute packages, it calls
CreateFile().
CreateFile() is problematic here, if path is a pipe, it will open
the pipe once then close, causing the other side to connect to the
wrong pipe. Skip stat() if the name is pipe and pretend there is one.