Create an ArrayExpr from an array of constants
Earlier we didn't bother converting an in list of constants into an ArrayExpr
when then number of elements was larger than array_expansion_threshold.

We still should create an ArrayExpr because it can be useful in deriving stats
on filters even if we are not deriving constraints on it.

Bump ORCA version to v3.47.0