Error if about to store an invalid segment number.
A user found a scenario where they had stored a -1 for a tuple's
segno. We'd like to be notified of this type of scenario because it
should not exist and we would like to track down the root
cause. Validations that throw errors will help us track the root cause
down.

(cherry picked from commit 19a0c80011735258e4a8ffc376983f67195021d0)