Inject faults by sending a libpq message
Faults can now be injected by sending a libpq message to postmaster, even if
the postmaster is acting as a mirror/standby.  The client must set
"gpconntype=fault" in libpq connection string in order to establish a fault
injection connection.  The fault details are sent as a separate message.  This
is similar to how FTS probe messages are sent.

A nice consequence is tests no longer have to rely on NOTICE messages emitted
by gp_inject_fault().  The new fault injection function returns a text datum,
which is easy to record and validate in answer files.

The old fault injection interface is left unchanged for now and the new
interface is added as "gp_inject_fault2()".  Replacing the old interface and
updating all tests and answer files can be done as a follow up patch.

Reviewed by: Jimmy Yih <jyih@pivotal.io>

(cherry picked from commit 89791f142e47512f47ce2d9f2eae36db9ea851e8)