Refactor making databases 2PC and correct ALTER DATABASE SET TABLESPACE
High-level details:
1) Tear out existing 2PC infrastructure for databases
2) Introduce new structures (in-mem, on-disk) for making databases 2PC
3) Make sure ALTER DATABASE SET TABLESPACE (ADST) works correctly such
that no orphaned files remain in the source/target tablespace
directories in cases of commit/abort.
4) We add the dboid dir to pendingDbDeletes for abort after copying
dboid dir in movedb. This ensures that the dboid dir is cleaned up from
the target tablespace when we have a failure during the course of
execution of the ALTER.
5) Re-engineer tests to assert filesystem changes and accommodate
multiple failure scenarios during the course of a 2PC commit/abort.
6) Introduce a new generic mechanism for making mirrors catch up. This
is done by inserting a XLOG_XACT_NOOP record and then waiting for the
mirror to replay the record with gp_inject_fault2. The premise is that
if the mirror has replayed this latest record, it would have replayed
everything before the NOOP record. We have introduced a UDF and a fault
injection point to make this possible.
7) Update pg_xlogdump to dump out pending DB deletes on commit/abort.

Refactoring Notes:
1) The xlog insert for database drops in `DropDatabaseDirectory`
is no longer required since we are already WALing database drops
elsewhere.
2) `xact_get_distributed_info_from_commit` populated variables that were
useful in the context of its only caller `xact_redo_distributed_commit`.
If we keep it around there would be unnecessary code duplication -> thus
deleting it.
3) Use DropRelationFiles() instead of duplicating logic in
xact_redo_distributed_commit.
4) Refactor session-level lock release for the QD during movedb().
5) smgr is a layer that pertains only to relations. So, we drop the
smgr prefix for pending db delete functions.
6) DropDatabaseDirectories() pertains to dropping dboid dirs at the
filesystem level. This means that it belongs inside storage.h/c.
DropDatabaseDirectories forms the visible interface as contrasted to
DropDatabaseDirectory. So moved the latter and made it static.
7) Rename DatabaseDropStorage to ScheduleDbDirDelete to call out its
intent.
8) Extract pending db deletes and dboid dir removal to dedicated file
9) Tests for scenarios with and without faults now live in one
consolidated file.

Co-authored-by: Taylor Vesely <tvesely@pivotal.io>
Co-authored-by: Adam Berlin <aberlin@pivotal.io>
Co-authored-by: Jesse Zhang <jzhang@pivotal.io>
(cherry picked from commit b0208894eb5ca72e9e621f964d4d313b065f15d5)